test(api-db): move db-specific tests from api-core
Move expected_machine and machine_interface database tests into api-db
module test files, preserving the existing assertions and comments while
adding local setup for static assignment segment coverage.
