+++ crates/api-db/src/expected_machine.rs
@@ -448,3 +448,6 @@ pub async fn update(txn: &mut PgConnection, machine: &ExpectedMachine) -> Databa
     }
     Ok(())
 }
+
+#[cfg(test)]
+mod tests;
