+++ crates/api-db/src/machine_interface.rs
@@ -95,6 +95,9 @@ impl ColumnInfo<'_> for IdColumn {
     }
 }
 
+#[cfg(test)]
+mod tests;
+
 #[derive(Clone, Copy)]
 pub struct MacAddressColumn;
 impl ColumnInfo<'_> for MacAddressColumn {
